*IMPORTANT* seats boarding method finalized
The passenger is able to go through and find it's own seat. The logic is that within the boarding group color passenger is in, they try to go to the furthest seat(the biggest x value) when there is still seats to choose from within those x values the method randomly picks a seat. The collide method not implemented yet but this is huge